Search fix

+++ Reto8/app/src/main/java/db/CompanyOperations.java
@@ -61,7 +61,7 @@ public class CompanyOperations {
         Cursor cursor = database.query(CompanyDBHandler.TABLE_COMPANIES, allColumns, CompanyDBHandler.COLUMN_ID + "=?", new String[]{String.valueOf(id)}, null, null, null, null);
         if (cursor != null)
             cursor.moveToFirst();
-        Company e = new Company(Long.parseLong(cursor.getString(0)), cursor.getString(1), cursor.getString(2), cursor.getInt(3), cursor.getString(4), cursor.getString(5), cursor.getString(6));
+        Company e = new Company(Long.parseLong(cursor.getString(0)), cursor.getString(1), cursor.getString(2), cursor.getLong(3), cursor.getString(4), cursor.getString(5), cursor.getString(6));
         return e;
     }
 
@@ -74,7 +74,7 @@ public class CompanyOperations {
                 company.setComId(cursor.getLong(cursor.getColumnIndex(CompanyDBHandler.COLUMN_ID)));
                 company.setName(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_NAME)));
                 company.setUrl(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_URL)));
-                company.setTelephone(cursor.getInt(cursor.getColumnIndex(CompanyDBHandler.COLUMN_TELEPHONE)));
+                company.setTelephone(cursor.getLong(cursor.getColumnIndex(CompanyDBHandler.COLUMN_TELEPHONE)));
                 company.setEmail(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_EMAIL)));
                 company.setServices(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_SERVICES)));
                 company.setType(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_TYPE)));
@@ -93,7 +93,7 @@ public class CompanyOperations {
                 company.setComId(cursor.getLong(cursor.getColumnIndex(CompanyDBHandler.COLUMN_ID)));
                 company.setName(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_NAME)));
                 company.setUrl(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_URL)));
-                company.setTelephone(cursor.getInt(cursor.getColumnIndex(CompanyDBHandler.COLUMN_TELEPHONE)));
+                company.setTelephone(cursor.getLong(cursor.getColumnIndex(CompanyDBHandler.COLUMN_TELEPHONE)));
                 company.setEmail(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_EMAIL)));
                 company.setServices(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_SERVICES)));
                 company.setType(cursor.getString(cursor.getColumnIndex(CompanyDBHandler.COLUMN_TYPE)));
